split main into files
