+++ main.py
@@ -1,28 +1,19 @@
-# This is a sample Python script.
+from set_parameters import set_simulation_parameters
+from concurrent.futures import ProcessPoolExecutor
+from multiple_photons import simulate_one_gc
+from itertools import repeat
+from plot_them import plot_GC_vs_angles_plot
 
-# Press ⌃R to execute it or replace it with your code.
-# Press Double ⇧ to search everywhere for classes, files, tool windows, actions, and settings.
+# -----------------------------
+# RUN SIMULATION BASED ON NUMBER OF PHOTONS AND GLUCOSE LEVEL INPUTS, WRITE OUT TO A GRAPH
+# -----------------------------
+if __name__ == "__main__":
 
+    n_cores, n_photons, GC_a = set_simulation_parameters()
 
-def print_hi(name):
-    # Use a breakpoint in the code line below to debug your script.
-    print(f'Hi, {name}')  # Press ⌘F8 to toggle the breakpoint.
+    angles = []
 
+    with ProcessPoolExecutor(max_workers=n_cores) as pool:
+        angles = list(pool.map(simulate_one_gc, GC_a, repeat(n_photons)))
 
-# Press the green button in the gutter to run the script.
-if __name__ == '__main__':
-
-    start = 0
-    alive = True
-
-    while alive:
-        end = start
-        start += 1
-        if start == 10:
-            alive = False
-
-        print ("start", start)
-        print ("end", end)
-
-
-# See PyCharm help at https://www.jetbrains.com/help/pycharm/
+    plot_GC_vs_angles_plot(GC_a, angles)
